Changed return value for RandUni from int to double.
I accidentally declared the return variable for RandUni as an int, resulting in all calls returning 0. This changes the variable to a double.

+++ rands.c
@@ -55,7 +55,7 @@ void RandSeed(signed long seed, pcg32_random_t* pcg_rng) {
 */
 double RandUni(pcg32_random_t* pcg_rng) {
 
-  int number;
+  double number;
 
 #ifdef RSOILWAT
